ENH Refactor FileIDHelper classes to reduce code duplication
